:art: fix https://github.com/siyuan-note/siyuan/issues/9004
